NOW WE CAN RUN JEST TESTS FROM GRUNT
We installed grunt-run and Jest with:
npm install --save-dev grunt-run
npm install --save-dev Jest

We added an addNumbers function to index.js and added a new
specs directory and created an index.spec.js file with a simple test
for addNumbers.

We also added "npm Jest" to our package.json test script. We updated
Gruntfile.js with a configuration for grunt-run so that it invokes
npm test. Now we can type npm test, or grunt test, and the latter
will read the test script in package.json and invoke it.